fix: #7 - 비슷한 컨트롤러들 합치기, 컨트롤러&서비스 클래스명에 Query 추가
