+++ src/main/java/com/threego/algo/studyrecruit/query/service/StudyRecruitPostQueryService.java
@@ -4,7 +4,7 @@ import com.threego.algo.studyrecruit.query.dto.*;
 
 import java.util.List;
 
-public interface StudyRecruitPostService {
+public interface StudyRecruitPostQueryService {
     List<StudyRecruitPostDTO> findStudyRecruitList(StudyRecruitSearchDTO searchDto);
 
     StudyRecruitDetailDTO findStudyRecruitDetail(Long postId);
